dev: refactored api service
