--- api/app/README.md
@@ -1,17 +1,0 @@
-# TherapEase API
-
-The TherapEase API is a proxy for the underlying services of TherapEase.
-
-## Requirements
-
-- Python 3.10+
-- Fast API
-- Pydantic
-
-## Service
-
-The following services are being proxied by this service:
-
-- Geocoding
-- Full Text Search
-- Facilities CRUD service (TODO)
