--- api/app/build.sh
@@ -1,2 +1,0 @@
-#! /bin/bash
-docker build -t therapease_api .
