--- api/app/tests/conftest.py
@@ -1,34 +1,0 @@
-from typing import Any
-from typing import Generator
-
-import pytest
-from fastapi import FastAPI
-from fastapi.testclient import TestClient
-
-import sys
-import os
-sys.path.append(os.path.dirname(os.path.dirname(os.path.abspath(__file__)))) 
-#this is to include backend dir in sys.path so that we can import from db,main.py
-
-from routers import facilities, search
-
-
-def start_application():
-    app = FastAPI()
-    app.include_router(search)
-    app.include_router(facilities)
-
-    return app
-
-
-
-@pytest.fixture(scope="function")
-def client(
-    app: FastAPI
-) -> Generator[TestClient, Any, None]:
-    """
-    Create a new FastAPI TestClient that uses the `db_session` fixture to override
-    the `get_db` dependency that is injected into routes.
-    """
-    with TestClient(app) as client:
-        yield client
