+++ api/src/app/libs/search/full_text.py
@@ -1,6 +1,7 @@
-import requests
-import dependencies
+from app.config import get_settings
 from .base import BaseSearch
+
+settings = get_settings()
 
 
 class FullTextSearch(BaseSearch):
@@ -8,4 +9,4 @@ class FullTextSearch(BaseSearch):
         """
         Initialize search service.
         """
-        super().__init__(dependencies.SEARCH_URL, dependencies.SEARCH_TOKEN, "search")
+        super().__init__(settings.SEARCH_URL, settings.SEARCH_TOKEN, "search")
