+++ api/src/requirements.txt
@@ -10,3 +10,7 @@ redis==4.6.0
 asyncpg==0.28.0
 fastapi-users==12.1.1
 fastapi-users-db-sqlalchemy==6.0.1
+pydantic==2.3.0
+pydantic-settings==2.0.3
+SQLAlchemy==2.0.20
+databases==0.6.2
