+++ api/src/tests/conftest.py
@@ -1,0 +1,10 @@
+import pytest
+from starlette.testclient import TestClient
+
+from app.main import app
+
+
+@pytest.fixture(scope="module")
+def test_app():
+    client = TestClient(app)
+    yield client  # testing happens here
